fix: The SQL component's Model is pulled away
