+++ ui/src/models/dataAnalysis.ts
@@ -2,11 +2,14 @@ import useRequest from "@/hooks/useRequest/useRequest";
 import systemApi from "@/services/systemSetting";
 import dataLogsApi from "@/services/dataLogs";
 import useRealTimeTraffic from "@/models/dataanalysis/useRealTimeTraffic";
-import useTemporaryQuery from "@/models/dataanalysis/useTemporaryQuery";
+import useTemporaryQuery, {
+  openNodeDataType,
+} from "@/models/dataanalysis/useTemporaryQuery";
 import useDataSourceManage from "@/models/dataanalysis/useDataSourceManage";
 import { InstanceType } from "@/services/systemSetting";
 import { useEffect, useState } from "react";
 import useWorkflow from "@/models/dataanalysis/useWorkflow";
+import dataAnalysis from "@/services/temporaryQuery";
 
 const DataAnalysis = () => {
   const [navKey, setNavKey] = useState<string>();
@@ -14,11 +17,34 @@ const DataAnalysis = () => {
   const [currentInstances, setCurrentInstances] = useState<number>();
   const [sqlQueryResults, setSqlQueryResults] = useState<any>();
   const [visibleSqlQuery, setVisibleSqlQuery] = useState<boolean>(false);
+  // 打开的文件节点id
+  const [openNodeId, setOpenNodeId] = useState<number>();
+  // 打开的文件节点父级id
+  const [openNodeParentId, setOpenNodeParentId] = useState<number>();
+  const [openNodeData, setOpenNodeData] = useState<openNodeDataType>();
+  // 节点修改后的value
+  const [folderContent, setFolderContent] = useState<string>("");
 
   const realTimeTraffic = useRealTimeTraffic();
   const temporaryQuery = useTemporaryQuery();
   const workflow = useWorkflow();
   const dataSourceManage = useDataSourceManage();
+
+  const changeOpenNodeId = (id: number) => {
+    setOpenNodeId(id);
+  };
+
+  const changeOpenNodeParentId = (parentId: number) => {
+    setOpenNodeParentId(parentId);
+  };
+
+  const changeOpenNodeData = (value: any) => {
+    setOpenNodeData(value);
+  };
+
+  const changeFolderContent = (str: string) => {
+    setFolderContent(str);
+  };
 
   const onChangeNavKey = (key: string) => {
     setNavKey(key);
@@ -48,12 +74,64 @@ const DataAnalysis = () => {
     loadingText: false,
   });
 
+  const doGetNodeInfo = useRequest(dataAnalysis.getNodeInfo, {
+    loadingText: false,
+  });
+
+  // Node
+  const doCreatedNode = useRequest(dataAnalysis.createdNode, {
+    loadingText: false,
+  });
+
+  const doUpdateNode = useRequest(dataAnalysis.updateNode, {
+    loadingText: false,
+  });
+
+  const doDeleteNode = useRequest(dataAnalysis.deleteNode, {
+    loadingText: false,
+  });
+
+  const doLockNode = useRequest(dataAnalysis.lockNode, {
+    loadingText: false,
+  });
+
+  const doUnLockNode = useRequest(dataAnalysis.unLockNode, {
+    loadingText: false,
+  });
+
+  const doRunCodekNode = useRequest(dataAnalysis.runCodekNode, {
+    loadingText: {
+      loading: "运行中",
+      done: "运行成功",
+    },
+  });
+
   useEffect(() => {
-    temporaryQuery.changeOpenNodeId(0);
-    temporaryQuery.changeOpenNodeParentId(0);
-    temporaryQuery.changeOpenNodeData(undefined);
-    temporaryQuery.changeFolderContent("");
+    changeOpenNodeId(0);
+    changeOpenNodeParentId(0);
+    changeOpenNodeData(undefined);
+    changeFolderContent("");
   }, [navKey]);
+
+  // 获取文件信息
+  const onGetFolderList = () => {
+    openNodeId &&
+      doGetNodeInfo.run(openNodeId).then((res: any) => {
+        if (res.code == 0) {
+          setOpenNodeData(res.data);
+          changeFolderContent(res.data.content);
+        }
+      });
+  };
+
+  // 是否修改
+  const isUpdateStateFun = () => {
+    return folderContent !== openNodeData?.content;
+  };
+
+  useEffect(() => {
+    onGetFolderList();
+  }, [openNodeId]);
 
   return {
     instances,
@@ -68,14 +146,38 @@ const DataAnalysis = () => {
     changeSqlQueryResults,
     changeVisibleSqlQuery,
 
+    folderContent,
+    changeFolderContent,
+
+    openNodeData,
+    changeOpenNodeData,
+
+    openNodeId,
+    changeOpenNodeId,
+
+    openNodeParentId,
+    changeOpenNodeParentId,
+    isUpdateStateFun,
+
+    onGetFolderList,
+
     doGetInstance,
     doGetDatabase,
     doGetTables,
+    doGetNodeInfo,
 
     realTimeTraffic,
     temporaryQuery,
     workflow,
     dataSourceManage,
+
+    // node
+    doCreatedNode,
+    doUpdateNode,
+    doDeleteNode,
+    doLockNode,
+    doUnLockNode,
+    doRunCodekNode,
   };
 };
