+++ ui/src/pages/DataAnalysis/components/SQLEditor/EditorContent/index.tsx
@@ -3,9 +3,9 @@ import MonacoEditor from "react-monaco-editor";
 import { useModel } from "umi";
 
 const EditorContent = () => {
-  const { temporaryQuery } = useModel("dataAnalysis");
+  const { openNodeData, changeFolderContent, folderContent } =
+    useModel("dataAnalysis");
   const { currentUser } = useModel("@@initialState").initialState || {};
-  const { openNodeData, changeFolderContent, folderContent } = temporaryQuery;
 
   const onChangeFolderContent = (value: string) => {
     changeFolderContent(value);
